Add IP address to the memcached script
We want to take the memcached stats with an IP address in graphite.

We will no longer be running a single global instance of memcached. That
means we need a way to distinguish the running instances. Using the
instance's IP address is a simple distinction.

Not that we "clean" the IP address to use dashes, no dots. Dots are the
separator for graphite stats.